refactor: move table service interfaces to Contracts/
Move Sortable, Filterable, Searchable, Renderable and SupportsPagination
interfaces from src/Table/Services/ to src/Table/Contracts/, alongside
the existing Cells/, Filter/ and Pagination/ contracts. Services/ is now
home to concrete classes only (BaseTableProvider,
BasePaginatedTableProvider, TableProviderInterface).
