+++ src/Table/Contracts/Filterable.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Patrikjak\Utils\Table\Services;
+namespace Patrikjak\Utils\Table\Contracts;
 
 use Patrikjak\Utils\Common\ValueObjects\Filter\FilterCriteria;
 use Patrikjak\Utils\Table\ValueObjects\Filter\Definitions\FilterableColumn;
