+++ src/Table/Contracts/Searchable.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Patrikjak\Utils\Table\Services;
+namespace Patrikjak\Utils\Table\Contracts;
 
 interface Searchable extends Renderable
 {
