+++ src/Table/Services/BasePaginatedTableProvider.php
@@ -6,6 +6,7 @@ namespace Patrikjak\Utils\Table\Services;
 
 use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\Blade;
+use Patrikjak\Utils\Table\Contracts\SupportsPagination;
 use Patrikjak\Utils\Table\Dto\Pagination\Paginator as TablePaginator;
 use Patrikjak\Utils\Table\Dto\Pagination\Settings;
 use Patrikjak\Utils\Table\Dto\Parameters;
