+++ src/Table/Services/BaseTableProvider.php
@@ -9,6 +9,10 @@ use Illuminate\Contracts\Container\BindingResolutionException;
 use Illuminate\Support\Facades\Blade;
 use Patrikjak\Utils\Common\ValueObjects\Filter\FilterCriteria;
 use Patrikjak\Utils\Common\ValueObjects\Sort\SortCriteria;
+use Patrikjak\Utils\Table\Contracts\Filterable;
+use Patrikjak\Utils\Table\Contracts\Searchable;
+use Patrikjak\Utils\Table\Contracts\Sortable;
+use Patrikjak\Utils\Table\Contracts\SupportsPagination;
 use Patrikjak\Utils\Table\Dto\Filter\Settings as FilterSettings;
 use Patrikjak\Utils\Table\Dto\Parameters;
 use Patrikjak\Utils\Table\Dto\Search\Settings as SearchSettings;
